# Perf: used to perf local hardware capacity