Most of users always say why my ceph cluster is slower than others, especially in ssd env. The difference is related to hardware, os, runtime version and library version. But it's hard to compare CPU, memory, controller model and library version directly. This program help users know the detail numbers of the low-level performance measurements comparing to others. Such as condition variable wakeup latency, some system calls latencies. It would also be helpful for devs.

This program contains a collection of low-level performance measurements
for Ceph, which can be run either individually or altogether.  These
tests measure performance in a single stand-alone process, not in a cluster
with multiple servers.  Invoke the program like this:

    Perf test1 test2 ...

test1 and test2 are the names of individual performance measurements to
run.  If no test names are provided then all of the performance tests
are run.

The result like:

atomic_int_cmp              12.62ns    atomic_t::compare_and_swap
atomic_int_inc              11.99ns    atomic_t::inc
atomic_int_read             15.44ns    atomic_t::read
atomic_int_set               0.00ns    atomic_t::set
mutex_nonblock              66.32ns    Mutex lock/unlock (no blocking)
buffer_basic               186.40ns    buffer create, add one ptr, delete
buffer_encode_decode         1.93us    buffer create, encode/decode object, delete
buffer_basic_copy            1.02us    buffer create, copy small block, delete
buffer_copy                 46.12ns    copy out 2 small ptrs from buffer
buffer_extend_chunk        557.14ns    buffer add onto existing ptr
buffer_get_contiguous       16.57ns    Buffer::get_contiguous
buffer_iterator              1.15us    iterate over buffer with 5 ptrs
cond_ping_pong               7.16us    condition variable round-trip
div32                        9.07ns    32-bit integer division instruction
div64                       47.18ns    64-bit integer division instruction
function_call                3.03ns    Call a function that has not been inlined
eventcenter_poll           485.96ns    EventCenter::process_events (no timers or events)
eventcenter_dispatch         3.22us    EventCenter::dispatch_event_external latency
memcpy100                   22.09ns    Copy 100 bytes with memcpy
memcpy1000                 117.63ns    Copy 1000 bytes with memcpy
memcpy10000                545.39ns    Copy 10000 bytes with memcpy
ceph_str_hash_rjenkins      42.47ns    rjenkins hash on 16 byte of data
ceph_str_hash_rjenkins     430.07ns    rjenkins hash on 256 bytes of data
rdtsc                       14.83ns    Read the fine-grain cycle counter
cycles_to_seconds           12.58ns    Convert a rdtsc result to (double) seconds
cycles_to_seconds           17.57ns    Convert a rdtsc result to (uint64_t) nanoseconds
prefetch                    22.92ns    Prefetch instruction
serialize                  193.19ns    serialize instruction
lfence                       9.23ns    Lfence instruction
sfence                       3.01ns    Sfence instruction
spin_lock                   15.49ns    Acquire/release SpinLock
spawn_thread                18.86us    Start and stop a thread
perf_timer                 616.08ns    Start and stop a SafeTimer
throw_int                    7.12us    Throw an int
throw_int_call               7.14us    Throw an int in a function call
throw_exception              5.37us    Throw an Exception
throw_exception_call         7.15us    Throw an Exception in a function call
vector_push_pop              6.76ns    Push and pop a std::vector

Signed-off-by: Haomai Wang <haomaiwang@gmail.com>