@XinzeChi 2015-05-06
@yuyuyu101 , there are 3 libs which use the most cpu in my cluster, including lock, memory malloc, leveldb.
Maybe we could add the perf of malloc and the leveldb(or rocksdb) for this PR?