@yuyuyu101 2015-05-06
@XinzeChi This program aimed to provide with *low-level* cost measures. We won't do level-db api measure benchmark which is a complicated api. memory alloc bottleneck is in heavy workload env, we can't get a reasonable benchmark here.