@markhpc 2015-05-06
Hi Haomai, "Perf" might be a bad name since "perf" is the name of the linux kernel profiling tool.  Can we change it to something different?