@dachary 2015-05-13
:+1: could you add it to ceph.spec.in and debian/control so it gets included in the relevant packages ?