@dachary 2015-05-18
@yuyuyu101 could you run it to the gitbuilders to verify they are happy ? It looks good to me but we're never too careful with packaging ;-) When they pass, 
<pre>
Reviewed-by: Loic Dachary &lt;ldachary@redhat.com>
</pre>