@tchaikov 2015-05-26
@yuyuyu101 i rebased your changes onto master and pushed it to `ceph/wip-haomai-ceph-perf-local`[0]. turns out that you forgot to put the newly added source files into the dist tarball. see http://gitbuilder.sepia.ceph.com/gitbuilder-ceph-deb-precise-amd64-basic/#origin/wip-haomai-ceph-perf-local .

---
[0] guess that's what @dachary was expecting and worrying about.