@tchaikov 2015-05-27
@yuyuyu101 FYI, i pushed your changes to wip-haomai-ceph-perf-local, seems the gitbuilders are happy.