@gregsfortytwo 2015-06-01
@liewegas @yuyuyu101 The merge of this branch has made our tarball (i.e., make check) gitbuilders very unhappy. The i386 ones are complaining about "SSE instruction set not enabled" and amd64 is not passing unit tests (although I couldn't figure out why in a quick check).