@tchaikov 2015-06-01
@gregsfortytwo i prepared a patch on branch of `wip-fix-ceph-perf-local`, let's see if it fixes the build.