@tchaikov 2015-06-01
fixed posted to #4820