@yghannam 2015-07-30
Hi all,
This breaks builds on AArch64 (and possibly other arches). There are a lot of asm statements being used without checking for the current build architecture.